fix: adding endpointmanager package to resolve platfrom specific call to HNS
