+++ cns/client/client.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"net/url"
 	"time"
@@ -1022,31 +1021,6 @@ func (c *Client) GetHomeAz(ctx context.Context) (*cns.GetHomeAzResponse, error) 
 	}
 
 	return &getHomeAzResponse, nil
-}
-
-// GetEndpoint calls the EndpointHandlerAPI in CNS to retrieve the state of a given EndpointID
-func (c *Client) GetHNSEndpointID(ctx context.Context, endpointID string) ([]string, [][]net.IPNet, error) {
-	endpointResponse, err := c.GetEndpoint(ctx, endpointID)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "failed to read the endpoint from CNS state")
-	}
-	var hnsEndpointIDs []string
-	var ips [][]net.IPNet
-	for _, ipInfo := range endpointResponse.EndpointInfo.IfnameToIPMap {
-		hnsEndpointID := ipInfo.HnsEndpointID
-		// we need to get the HNSENdpoint via the IP address if the HNSEndpointID is not present in the statefile
-		if ipInfo.HnsEndpointID == "" {
-			if ipInfo.IPv4 != nil {
-				ips = append(ips, ipInfo.IPv4)
-			}
-			if ipInfo.IPv6 != nil {
-				ips = append(ips, ipInfo.IPv6)
-			}
-		} else {
-			hnsEndpointIDs = append(hnsEndpointIDs, hnsEndpointID)
-		}
-	}
-	return hnsEndpointIDs, ips, nil
 }
 
 // GetEndpoint calls the EndpointHandlerAPI in CNS to retrieve the state of a given EndpointID
