+++ cns/hnsclient/hnsclient_linux.go
@@ -2,7 +2,6 @@ package hnsclient
 
 import (
 	"fmt"
-	"net"
 
 	"github.com/Azure/azure-container-networking/cns"
 )
@@ -50,14 +49,4 @@ func CreateHostNCApipaEndpoint(
 func DeleteHostNCApipaEndpoint(
 	networkContainerID string) error {
 	return nil
-}
-
-// DeleteHNSEndpointbyID deletes the HNS endpoint
-// created for Satateless CNI Asynch delete
-func DeleteHNSEndpointbyID(_ string) error {
-	return nil
-}
-
-func GetHNSEndpointbyIP(_, _ []net.IPNet) (string, error) {
-	return "", nil
 }
